move canvas, gl, and shader init code to separate modules
