+++ js/index.js
@@ -1,116 +1,10 @@
 import 'https://code.jquery.com/jquery-3.5.1.min.js';
-import { EYE, LIGHTS, perspectiveView, RubiksCube, scramble, togglePerspective } from './modules/rubiks.js';
+import initCanvas from './modules/canvas.js';
+import initWebGL from './modules/gl.js';
+import { perspectiveView, RubiksCube, scramble, togglePerspective } from './modules/rubiks.js';
+import { initShader } from './modules/shader.js';
 
 const $ = window.$;
-
-function initCanvas() {
-    const $canvas = $('#glcanvas');
-    const canvas = $canvas[0];
-    canvas.width = canvas.clientWidth;
-    canvas.height = canvas.clientHeight;
-
-    window.onresize = function () {
-        canvas.width = canvas.clientWidth;
-        canvas.height = canvas.clientHeight;
-    };
-
-    return $canvas
-}
-
-function initWebGL(canvas) {
-    if (!window.WebGLRenderingContext) {
-        console.log("Your browser doesn't support WebGL.")
-        return null;
-    }
-
-    const gl = canvas.getContext('webgl', {preserveDrawingBuffer: true}) || canvas.getContext('experimental-webgl', {preserveDrawingBuffer: true});
-    if (!gl) {
-        console.log("Your browser supports WebGL, but initialization failed.");
-        return null;
-    }
-
-    gl.clearColor(1.0, 1.0, 1.0, 1.0);
-    gl.enable(gl.DEPTH_TEST);
-    gl.depthFunc(gl.LEQUAL);
-    gl.enable(gl.CULL_FACE);
-    gl.cullFace(gl.BACK);
-    gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
-
-    return gl;
-}
-
-function initShaders(gl) {
-    const fragmentShader = getShader(gl, 'fragmentShader');
-    const vertexShader = getShader(gl, 'vertexShader');
-    const shaderProgram = gl.createProgram();
-    gl.attachShader(shaderProgram, fragmentShader);
-    gl.attachShader(shaderProgram, vertexShader);
-    gl.linkProgram(shaderProgram);
-    if (!gl.getProgramParameter(shaderProgram, gl.LINK_STATUS)) {
-        console.log('Unable to initialize the shader program');
-    }
-    gl.useProgram(shaderProgram);
-
-    shaderProgram.vertexPosition = gl.getAttribLocation(shaderProgram, 'vertexPosition');
-    gl.enableVertexAttribArray(shaderProgram.vertexPosition);
-
-    shaderProgram.vertexNormal = gl.getAttribLocation(shaderProgram, 'vertexNormal');
-    gl.enableVertexAttribArray(shaderProgram.vertexNormal);
-
-    shaderProgram.eye = gl.getUniformLocation(shaderProgram, 'eye');
-    gl.uniform3fv(shaderProgram.eye, EYE);
-
-    for (let i = 0; i < LIGHTS.length; i++) {
-        const lightPosition = `lights[${i}].position`;
-        const lightIntensity = `lights[${i}].intensity`;
-        shaderProgram[lightPosition] = gl.getUniformLocation(shaderProgram, lightPosition);
-        shaderProgram[lightIntensity] = gl.getUniformLocation(shaderProgram, lightIntensity);
-        gl.uniform3fv(shaderProgram[lightPosition], LIGHTS[i].position);
-        gl.uniform1f(shaderProgram[lightIntensity], LIGHTS[i].intensity);
-    }
-
-    shaderProgram.lighting = gl.getUniformLocation(shaderProgram, 'lighting');
-    shaderProgram.ambient = gl.getUniformLocation(shaderProgram, 'ambient');
-    shaderProgram.diffuse = gl.getUniformLocation(shaderProgram, 'diffuse');
-    shaderProgram.specular = gl.getUniformLocation(shaderProgram, 'specular');
-    shaderProgram.specularExponent = gl.getUniformLocation(shaderProgram, 'specularExponent');
-
-    return shaderProgram;
-}
-
-function getShader(gl, id) {
-    const shaderScript = document.getElementById(id);
-    if (!shaderScript) {
-        return null;
-    }
-
-    let source = '';
-    let currentChild = shaderScript.firstChild;
-    while (currentChild) {
-        if (currentChild.nodeType === currentChild.TEXT_NODE) {
-            source += currentChild.textContent;
-        }
-        currentChild = currentChild.nextSibling;
-    }
-
-    let shader;
-    if (shaderScript.type === 'x-shader/x-fragment') {
-        shader = gl.createShader(gl.FRAGMENT_SHADER);
-    } else if (shaderScript.type === 'x-shader/x-vertex') {
-        shader = gl.createShader(gl.VERTEX_SHADER);
-    } else {
-        return null;
-    }
-
-    gl.shaderSource(shader, source);
-    gl.compileShader(shader);
-    if (!gl.getShaderParameter(shader, gl.COMPILE_STATUS)) {
-        console.log('An error occurred while compiling the shader: ' + gl.getShaderInfoLog(shader));
-        return null;
-    }
-
-    return shader;
-}
 
 // timestamp is a DOMHighResTimeStamp.
 // See https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame.
@@ -131,9 +25,9 @@ function drawScene(rubiksCube) {
 }
 
 $(document).ready(function() {
-    const $canvas = initCanvas();
+    const $canvas = initCanvas('#glcanvas');
     const gl = initWebGL($canvas[0]);
-    const shaderProgram = initShaders(gl);
+    const shaderProgram = initShader(gl);
     const pathname = location.pathname;
     const base = pathname.substring(0, pathname.lastIndexOf('/'));
