+++ js/modules/bounding-box.js
@@ -1,11 +1,11 @@
 import './gl-matrix-min.js';
+import { EYE } from './shader.js';
 
 // The bounding box is a 6x6x6 cube, centered at the origin, that inscribes the Rubik's cube.
 // Each face of the bounding box corresponds to a side of the Rubik's cube.
-function BoundingBox(gl, projectionMatrix, modelViewMatrix, eye) {
+function BoundingBox(gl, projectionMatrix, modelViewMatrix) {
     this.projectionMatrix = projectionMatrix;
     this.modelViewMatrix = modelViewMatrix;
-    this.eye = eye;
     this.planes = [
         new Plane([-3, -3, 3], [3, -3, 3], [-3, 3, 3]), // front
         new Plane([-3, -3, -3], [-3, 3, -3], [3, -3, -3]), // back
@@ -55,7 +55,7 @@ function BoundingBox(gl, projectionMatrix, modelViewMatrix, eye) {
                 //
                 // The bounding box is in model space (local coordinates). We transform
                 // it into world space when measuring the distance to the eye/camera to account for rotations.
-                const distance = glMatrix.vec3.distance(this.eye, glMatrix.vec3.fromValues(...worldPoint));
+                const distance = glMatrix.vec3.distance(EYE, glMatrix.vec3.fromValues(...worldPoint));
                 if (distance < minDistance) {
                     minDistance = distance;
                     intersection = {
