+++ js/modules/gl.js
@@ -1,0 +1,23 @@
+function initWebGL(canvas) {
+    if (!window.WebGLRenderingContext) {
+        console.log("Your browser doesn't support WebGL.")
+        return null;
+    }
+
+    const gl = canvas.getContext('webgl', {preserveDrawingBuffer: true}) || canvas.getContext('experimental-webgl', {preserveDrawingBuffer: true});
+    if (!gl) {
+        console.log("Your browser supports WebGL, but initialization failed.");
+        return null;
+    }
+
+    gl.clearColor(1.0, 1.0, 1.0, 1.0);
+    gl.enable(gl.DEPTH_TEST);
+    gl.depthFunc(gl.LEQUAL);
+    gl.enable(gl.CULL_FACE);
+    gl.cullFace(gl.BACK);
+    gl.clear(gl.COLOR_BUFFER_BIT | gl.DEPTH_BUFFER_BIT);
+
+    return gl;
+}
+
+export default initWebGL;
